+++ aospy/data_loader.py
@@ -146,11 +146,11 @@ def _prep_time_data(ds):
 
     Returns
     -------
-    Dataset, int
-        The processed Dataset and minimum year in the loaded data
+    Dataset, int, int
+        The processed Dataset and minimum and maximum years in the loaded data
     """
     ds = times.ensure_time_as_dim(ds)
-    ds, min_year = times.numpy_datetime_workaround_encode_cf(ds)
+    ds, min_year, max_year = times.numpy_datetime_workaround_encode_cf(ds)
     if internal_names.TIME_BOUNDS_STR in ds:
         ds = times.ensure_time_avg_has_cf_metadata(ds)
     else:
@@ -161,7 +161,7 @@ def _prep_time_data(ds):
     ds = xr.decode_cf(
         ds, decode_times=True, decode_coords=False, mask_and_scale=False
     )
-    return ds, min_year
+    return ds, min_year, max_year
 
 
 def _load_data_from_disk(file_set, preprocess_func=lambda ds: ds, **kwargs):
@@ -232,13 +232,13 @@ class DataLoader(object):
         ds = _load_data_from_disk(file_set, self.preprocess_func,
                                   start_date=start_date, end_date=end_date,
                                   time_offset=time_offset, **DataAttrs)
-        ds, min_year = _prep_time_data(ds)
+        ds, min_year, max_year = _prep_time_data(ds)
         ds = set_grid_attrs_as_coords(ds)
         da = _sel_var(ds, var)
         da = self._maybe_apply_time_shift(da, time_offset, **DataAttrs)
 
         start_date_xarray = times.numpy_datetime_range_workaround(
-            start_date, min_year)
+            start_date, min_year, max_year)
         end_date_xarray = start_date_xarray + (end_date - start_date)
         return times.sel_time(da, np.datetime64(start_date_xarray),
                               np.datetime64(end_date_xarray)).load()